Reorganise Wasm related CLIs in test framework
