+++ tools/test-framework/src/chain/cli/wasm/mod.rs
@@ -1,0 +1,2 @@
+pub mod contract;
+pub mod query;
