+++ tools/test-framework/src/chain/ext/async_icq.rs
@@ -1,7 +1,6 @@
 use crate::chain::cli::async_icq::{async_icq, update_oracle};
-use crate::chain::cli::query::{
-    instantiate_wasm_contract, query_wasm_list_code, query_wasm_list_contracts_by_code,
-};
+use crate::chain::cli::wasm::contract::instantiate_wasm_contract;
+use crate::chain::cli::wasm::query::{query_wasm_list_code, query_wasm_list_contracts_by_code};
 use crate::prelude::*;
 use crate::types::tagged::*;
